requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
cachetools==5.3.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==41.0.2
cvzone==1.5.6
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0.post2
fastjsonschema==2.18.0
filelock==3.12.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.4.0
gast==0.4.0
gevent==23.7.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==21.2.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
kombu==5.3.1
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
nbformat==5.9.2
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pigar==2.1.1
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.30.2
requests==2.30.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sounddevice==0.4.6
SQLAlchemy==2.0.14
start==0.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.147
urllib3==1.26.15
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
WTForms==3.0.1
xgboost==1.7.6
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
